% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{add_Module}
\alias{add_Module}
\title{Add a Gene Module Score without numeric tail}
\usage{
add_Module(obj, genelist)
}
\arguments{
\item{obj}{Seurat object to be scored}

\item{geneset}{Vector of gene symbols}
}
\value{
Seurat object with calculated module score as meta.data column
}
\description{
This function adds a gene module score to a Seurat object.
Built off the Seurat-native AddModuleScore() function. Unlike AddModuleScore(),
this function generates the new column with the exact string given.
}
