% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{clean_enrichR}
\alias{clean_enrichR}
\title{Clean the dirty data output of enrichr()}
\usage{
clean_enrichR(df, padj_cutoff = 0.05)
}
\arguments{
\item{df}{EnrichR output dataframe}
}
\value{
Dataframe with cleaner data and statistical cutoff
}
\description{
Clean the dirty data output of enrichr()
}
