% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{find_Correlations}
\alias{find_Correlations}
\title{Find correlations between a query and a geneset}
\usage{
find_Correlations(obj, query, subset = NULL)
}
\arguments{
\item{obj}{Seurat object}

\item{query}{Vector of genes to assess}

\item{subset}{Subset of total genes in obj to correlate against}
}
\description{
This function calculates the Pearson correlation between a query gene and
all detected genes in the Seurat object, returning a dataframe.
Specify a gene subset to improve efficiency.
}
