% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{percent_Nonzero}
\alias{percent_Nonzero}
\title{Get percent nonzero expressing cells in Seurat obj}
\usage{
percent_Nonzero(object, genes)
}
\arguments{
\item{obj}{Seurat object}

\item{geneset}{Vector of genes}
}
\description{
This function fetches expression data for a given geneset and
calculates the percentage of cells in the whole object with non-zero
expression levels for each gene.
}
