% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_composition}
\alias{plot_composition}
\title{Plot differences in cell composition}
\usage{
plot_composition(
  obj,
  group.by,
  split.by = NULL,
  as.pie = FALSE,
  group.levels = NULL,
  split.levels = NULL,
  label = FALSE,
  label.pct = FALSE
)
}
\arguments{
\item{obj}{Seurat object}

\item{group.by}{Meta.data column name to annotate}

\item{split.by}{Meta.data column name to split plots}

\item{as.pie}{Output plot as pie chart, defaults to FALSE}

\item{group.levels}{Vector of levels to order group.by with}

\item{split.levels}{Vector of levels to order split.by with}

\item{label}{Labels to be plotted on plot instead of in legend, defaults to FALSE}

\item{label.pct}{Label to be plotted with proportion, defaults to FALSE}
}
\description{
Plot differences in cell composition
}
