% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrichr}
\alias{plot_enrichr}
\title{Plot the output of the enrichR dataframe}
\usage{
plot_enrichr(df, ntop, fill_by = "Overlap", title = NULL, term_char = 100)
}
\arguments{
\item{df}{EnrichR output dataframe cleaned with clean_enrichR}

\item{ntop}{Number of top terms to plot}

\item{fill_by}{Variable to fill with}

\item{title}{Plot title}

\item{term_char}{Max number of characters before truncating term}
}
\value{
A ggplot2 plot of top enrichement hits
}
\description{
Plot the output of the enrichR dataframe
}
