% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_Nonzero}
\alias{plot_Nonzero}
\title{Plot percent of cells with any detectable expression}
\usage{
plot_Nonzero(
  obj,
  gene,
  x.margin = 5,
  y.margin = 0,
  title.margin = 5,
  title.size = 8,
  color = "blue"
)
}
\arguments{
\item{obj}{Seurat object}

\item{gene}{Single gene name}

\item{x.margin}{X margin size}

\item{y.margin}{Y margin size}

\item{title.margin}{Title margin from rectangle}

\item{title.size}{Title text size}

\item{color}{Color of percent fill}
}
\description{
Plot percent of cells with any detectable expression
}
