% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{read_GO}
\alias{read_GO}
\title{Get genes from a downloadable Gene Ontology text file (MGI)}
\usage{
read_GO(path)
}
\arguments{
\item{path}{File or directory path}
}
\value{
List of gene vectors
}
\description{
This function reads a MGI formated GO file and extracts the genes that
make up the gene ontology term.
}
