% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{theme_seurat}
\alias{theme_seurat}
\title{Generate ggplot2 themes for common Seurat plots}
\usage{
theme_seurat(plot_type = "base")
}
\arguments{
\item{plot_type}{One of "base", "dot", "dim", "vln", or "comp"}
}
\value{
ggplot2 theme
}
\description{
This function provides custom themes for several
common Seurat functions, making them presentation-
ready.
}
